category hover menu and search modal
